MusicDownloader/requirements.txt
# basic
vk-api>=11.8
requests>=2.22
bs4
# music related
pydub
mutagen
music-tag
streamlink
ffmpeg-python